+++ packages/examples/src/verifyPrepare.ts
@@ -1,3 +1,4 @@
 export * from 'monaco-editor-wrapper';
+export * from 'monaco-editor-wrapper/styles';
 import 'monaco-editor/esm/vs/basic-languages/javascript/javascript.contribution.js';
 import 'monaco-editor/esm/vs/language/typescript/monaco.contribution.js';