+++ packages/monaco-editor-wrapper/src/index.ts
@@ -78,5 +78,4 @@ export {
     EditorAppVscodeApi
 };
 
-export * from './helpers/css.js';
 export * from './utils.js';