+++ packages/monaco-editor-wrapper/src/utils.ts
@@ -23,3 +23,7 @@ export const createUrl = (config: WebSocketConfigOptions | WebSocketConfigOption
     }
     return buildUrl;
 };
+
+export const verifyUrlorCreateDataUrl = (input: string | URL) => {
+    return (input instanceof URL) ? input.href : new URL(`data:text/plain;base64,${btoa(input)}`).href;
+};